Issue #500 - Fixing a big image leak
This leak happens in the Properties view. If there is some color property in
the Properties view, we show image and label. The label show rbg
numbers, but we create a new image whenever property viewer requires
it (redisplay, refresh ...). You will get leak(s) when painting in
editor, when closing/opening the Properties view, dbl-click (select)
some figure in editor that has some color property...
It's possible to get "no more handles" relative quickly.
The image is created based on RBG property and it's created always
when the Property Sheet updates a color property (it even happens 2
(maybe more) times for the same property).
The fix uses image cache for rgb object. The cache is cleared after 200 images.